"Did a bunch of stuff with the non-Admin user like adding views ext"

+++ code/ngo_app/views.py
@@ -3,30 +3,26 @@ from django.contrib.auth.views import LoginView , LogoutView
 from django.urls import reverse_lazy
 from django.views.generic import *
 from django.views import View
-from .forms import UserDataForms
-from .models import UserData, AppUser ,Events
+from .models import EventRegistration,Events ,User , Group
 from .forms import UserDataForms
 from django.shortcuts import get_object_or_404 ,redirect
 from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
+from .forms import AddUserForm
+
 # Create your views here.
 class HomeView(LoginView, LogoutView):
     pass
 
 
-class BasicAppView(View):
-
-    def get(self,request):
-        form = UserDataForms()
-        return render(request, "base.html", {"form": form})
-
 
 class ListAll(ListView):
-    queryset = AppUser.objects.all()
-    template_name = "list.html"
+    queryset = User.objects.all()
+    template_name = "AdminTable.html"
 
 
 class UpdateStuff(UpdateView):
-    model = UserData
+    model = EventRegistration
     fields = ['first_name', "last_name", 'address_line1', 'address_line2', 'city', 'state_code', 'zip', 'country']
     template_name = 'base.html'
     pk_url_kwarg = 'pk'
@@ -36,33 +32,61 @@ class UpdateStuff(UpdateView):
         return render(request, "base.html", {'form': form})
 
 class UpdateUsers(UpdateView):
-    model = AppUser
-    fields = ['first_name', 'last_name', 'email','group']
+    model = User
+    fields = ['first_name', 'last_name', 'email']
     template_name = 'base.html'
     pk_url_kwarg = 'pk'
     slug_url_kwarg = 'first_name'
     slug_field = 'First Name'
     success_url = '/admin'
+    def get(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            return super().get(request, args, kwargs)
+        return redirect('/login')
+
+    def post(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            return super().post(request, args, kwargs)
+        return redirect('/login')
 
 class DelUser(DeleteView):
-    model = AppUser
+    model = User
 
     def get(self, request, *args, **kwargs):
-        try:
-            pkkey = kwargs['pk']
-        except:
-            return HttpResponse(status=500)
+        if request.user.is_authenticated:
+            try:
+                pkkey = kwargs['pk']
+            except:
+                return HttpResponse(status=500)
 
-        user = get_object_or_404(AppUser, pk=pkkey)
-        user.delete()
-        return redirect('/admin/')
+            user = get_object_or_404(User, pk=pkkey)
+            user.delete()
+            return redirect('/admin/')
 
 
-class AddUser(CreateView):
-    model = AppUser
-    fields = ['first_name', 'last_name', 'email', 'group']
+
+class AddUser(View):
+    model = User
+    fields = ['username','first_name', 'last_name', 'email', 'password']
     template_name = 'base.html'
     success_url = '/admin'
+
+    def get(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            form = AddUserForm
+            return render(request, 'base.html', {'form':form})
+        return redirect('/login')
+
+    def post(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            form = AddUserForm(request.POST)
+            super(AddUserForm, form).clean()
+            ustobj = form.save()
+            ustobj.groups.set([get_object_or_404(Group, name="User")])
+            ustobj.save()
+            return redirect("/admin/")
+        return redirect('/login')
+
 
 
 class AddEvent(CreateView):
@@ -76,17 +100,41 @@ class UpdateEvent(UpdateView):
     fields = ['name', 'type', 'status']
     template_name = 'base.html'
     pk_url_kwarg = 'pk'
+    success_url = '/admin'
 
 
 class DelEvent(DeleteView):
     model = Events
 
     def get(self, request, *args, **kwargs):
-        try:
-            pkkey = kwargs['pk']
-        except:
-            return HttpResponse(status=500)
+        if request.user.is_authenticated:
+            try:
+                pkkey = kwargs['pk']
+            except:
+                return HttpResponse(status=500)
 
-        evnt = get_object_or_404(Events, pk=pkkey)
-        evnt.delete()
-        return redirect('/admin/')
+            evnt = get_object_or_404(Events, pk=pkkey)
+            evnt.delete()
+            return redirect('/admin/')
+
+class EvenRegistrationView(CreateView):
+    model = EventRegistration
+    fields = ['first_name']#, 'last_name', 'cma', 'phone', 'email', 'address_line1', 'address_line2', 'city','state_code', 'zip', 'country', 'urbanization']
+    template_name = 'base.html'
+    success_url = '/admin/'
+
+    def post(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            #registration =
+            return super().post(request, args, kwargs)
+
+
+
+
+class ListCArtView(ListView):
+    queryset = Events.objects.all()
+    template_name = 'CartTable.html'
+
+
+
+
